+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Improve support for recursive type aliases
 - Correctly handle type aliases and other types with fallbacks in more places
 - Fix edge case in `TypeIs` type narrowing with tuple types
 - Rewrite the implementation of assignability to be more in line with the typing